backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Cache & Message Queue
redis==5.2.1
celery==5.4.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.10.2
pydantic-settings==2.6.1

# Machine Learning & NLP
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1
numpy==1.26.4
scikit-learn==1.5.2

# Monitoring & Observability
prometheus-client==0.21.0
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0

# WebSocket & Real-time
websockets==14.1
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
httpx==0.28.1
